refactor: pagify into util function
