+++ src/commands/utility/search.ts
@@ -1,6 +1,7 @@
 import Command from '../../lib/structures/Command';
 import { TypicalGuildMessage } from '../../lib/types/typicalbot';
 import { Modes } from '../../lib/utils/constants';
+import { pagify } from '../../lib/utils/util';
 
 const regex = /(\S+)(?:\s+(\d+))?/i;
 
@@ -29,7 +30,7 @@ export default class extends Command {
         if (!list.size)
             return message.reply(message.translate('utility/search:NONE', { query }));
 
-        const content = this.client.helpers.pagify.execute(message, list.map((m) =>
+        const content = pagify(message, list.map((m) =>
             `${`${m.user.username}${
                 m.nickname ? ` (${m.nickname})` : ''
             }`.padEnd(40)}: ${m.id}`), page);
