--- src/functions/pagify.ts
@@ -1,25 +1,0 @@
-import Function from '../lib/structures/Function';
-import { TypicalGuildMessage } from '../lib/types/typicalbot';
-
-export default class Pagify extends Function {
-    execute(message: TypicalGuildMessage, list: string[], page = 1) {
-        const listSize = list.length;
-        const pageCount = Math.ceil(listSize / 10);
-
-        page = page > pageCount ? 0 : page - 1;
-
-        const currentPage = list.splice(page * 10, 10);
-
-        const pageContent = currentPage
-            .map((item, index) =>
-                `• ${String(index + 1 + 10 * page).padStart(String(10 + 10 * page).length)}: ${item}`)
-            .join('\n');
-
-        return message.translate('misc:PAGIFY', {
-            page: page + 1,
-            pages: pageCount,
-            total: listSize.toLocaleString(),
-            content: pageContent
-        });
-    }
-}
